clicks now handled through itemTouchListener

+++ app/src/main/java/com/lectricas/curriencies/CurrencyItem.kt
@@ -7,7 +7,7 @@ data class CurrencyItem(
     val firstItem: Boolean = false
 ) {
     companion object {
-        fun newMultiplied(item: CurrencyItem, amount: Int): CurrencyItem {
+        fun newMultiplied(item: CurrencyItem, amount: Double): CurrencyItem {
             return CurrencyItem(
                 item.id,
                 item.multiplier,
